pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1"]
build-backend = "setuptools.build_meta"

[project]
name = "core-graph"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "psycopg[binary]>=3.1",
  "nats-py>=2.7",
  "redis>=5.0",
  "pydantic>=2.5",
  "mcp>=1.0",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "jellyfish>=1.0",
  "httpx>=0.27",
]
dev = [
  "ruff>=0.4",
  "yamllint>=1.35",
]

[tool.setuptools.packages.find]
include = ["ingest", "ingest.*", "api", "api.*", "evidence", "evidence.*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires running Docker stack (deselect with '-m not integration')",
]
